# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Shahebbazar API
  version: 1.1.0-sprint2
  summary: Rajshahi local business directory — Express API.
  description: |
    The Express API behind shahebbazar.com. The Next.js frontend calls it
    from the server when rendering pages; the provider dashboard, the
    analytics beacons and image uploads call it from the browser.

    **Browse it:** run the API and open `http://localhost:4000/api/docs`.
    **Postman:** File → Import → this file (`docs/openapi.yaml`).

    **Authentication is not built yet.** Every provider and admin
    endpoint is marked `TODO(auth)` in the code and is currently open.
    The customer (`/api/me/...`) endpoints act as the first active
    customer account until sign-in exists.
    They must not be deployed publicly until sessions and role checks
    exist. The provider endpoints take the business id in the path for
    now; with sign-in, it will come from the session.

    **Conventions**
    - All public responses contain approved records only.
    - `NUMERIC` and `COUNT(*)` values that PostgreSQL returns as big
      numbers arrive as **strings** (e.g. `"rating": "4.5"`); convert
      before doing arithmetic.
    - Stored image paths are relative (`uploads/…`) and are served by this
      API at `/uploads/…`.
    - Errors are `{ "error": "…" }`; form endpoints add `errors` keyed by
      field.
  contact:
    name: COS40005 Group 23
# No authentication exists yet, so no security scheme is declared. When
...
security: []
servers:
...
paths:
...
  /api/vendors/{vendorId}/dashboard:
    get:
      operationId: getVendorsByVendorIdDashboard
      tags: [Provider]
      summary: Everything the provider dashboard shows
      description: Works for a business in any status. Periods are whole days in Asia/Dhaka; the current period includes today.
      parameters:
        - $ref: "#/components/parameters/VendorId"
        - { name: days, in: query, schema: { type: integer, enum: [7, 30, 90], default: 30 }, description: "Other values fall back to 30." }
      responses:
        "200":
          description: Stat cards, chart series and listing preview.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VendorDashboard" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/vendors/{vendorId}/summary:
    get:
      operationId: getVendorsByVendorIdSummary
      tags: [Provider]
      summary: Name, slug and status of one business
      parameters:
        - $ref: "#/components/parameters/VendorId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VendorSummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  parameters:
...
    VendorId:
      name: vendorId
      in: path
      required: true
      description: TODO(auth) — will come from the session.
      schema: { type: integer, minimum: 1 }
      example: 13
...
  responses:
...
    BadRequest:
      description: Invalid input.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Not found (for public endpoints, also when not approved).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        detail: { type: string }
        hint: { type: string }
...
    NumericString:
      type: string
      pattern: '^-?\d+(\.\d+)?$'
      description: A PostgreSQL NUMERIC or BIGINT, as a string.
    NullableString: { type: [string, "null"] }
    VendorStatus: { type: string, enum: [draft, pending, approved, rejected, suspended] }
    ListingStatus: { type: string, enum: [draft, pending, approved, rejected, archived] }
    OpenState:
      type: [string, "null"]
      enum: [open, closed, null]
      description: Null means opening hours are not recorded, not that the business is closed.
...
    VendorSummary:
      type: object
      required: [vendor_id, vendor_name, vendor_slug, vendor_status]
      properties:
        vendor_id: { type: integer }
        vendor_name: { type: string }
        vendor_slug: { type: string }
        vendor_status: { $ref: "#/components/schemas/VendorStatus" }
    PeriodCount:
      type: object
      required: [current, previous]
      properties: { current: { type: integer }, previous: { type: integer } }
    VendorDashboard:
      type: object
      required: [vendor, days, stats, series, listings]
      properties:
        vendor:
          type: object
          required: [vendor_id, vendor_slug, vendor_name, vendor_phone, vendor_status, is_verified, vendor_created_at, open_state]
          properties:
            vendor_id: { type: integer }
            vendor_slug: { type: string }
            vendor_name: { type: string }
            vendor_description: { $ref: "#/components/schemas/NullableString" }
            vendor_phone: { type: string }
            vendor_email: { $ref: "#/components/schemas/NullableString" }
            vendor_address: { $ref: "#/components/schemas/NullableString" }
            vendor_logo_url: { $ref: "#/components/schemas/NullableString" }
            vendor_cover_url: { $ref: "#/components/schemas/NullableString" }
            vendor_status: { $ref: "#/components/schemas/VendorStatus" }
            vendor_rejection_reason: { $ref: "#/components/schemas/NullableString" }
            is_verified: { type: boolean }
            vendor_created_at: { type: string, format: date-time }
            area_name: { $ref: "#/components/schemas/NullableString" }
            primary_category: { $ref: "#/components/schemas/NullableString" }
            open_state: { $ref: "#/components/schemas/OpenState" }
            close_time_today: { $ref: "#/components/schemas/NullableString" }
        days: { type: integer, enum: [7, 30, 90] }
        stats:
          type: object
          required: [profileViews, searchAppearances, listings, rating]
          properties:
            profileViews: { $ref: "#/components/schemas/PeriodCount" }
            searchAppearances: { $ref: "#/components/schemas/PeriodCount" }
            listings:
              type: object
              required: [active, addedInPeriod, pending, total]
              properties:
                active: { type: integer }
                addedInPeriod: { type: integer }
                pending: { type: integer }
                total: { type: integer }
            rating:
              type: object
              required: [average, count]
              properties: { average: { type: [number, "null"] }, count: { type: integer } }
        series:
          type: array
          description: One row per day, oldest first, zero-filled. The last row is today (partial).
          items:
            type: object
            required: [day, views, appearances]
            properties:
              day: { type: string, format: date }
              views: { type: integer }
              appearances: { type: integer }
        listings:
          type: array
          maxItems: 3
          items:
            type: object
            required: [listing_id, listing_title, listing_status]
            properties:
              listing_id: { type: integer }
              listing_title: { type: string }
              listing_price: { oneOf: [{ $ref: "#/components/schemas/NumericString" }, { type: "null" }] }
              listing_price_max: { oneOf: [{ $ref: "#/components/schemas/NumericString" }, { type: "null" }] }
              listing_price_unit: { $ref: "#/components/schemas/NullableString" }
              listing_status: { $ref: "#/components/schemas/ListingStatus" }
              category_name: { $ref: "#/components/schemas/NullableString" }
              photo_url: { $ref: "#/components/schemas/NullableString" }
...